Migrating from main to master
The program is to be migrated from main branch on git to master. The
main branch has served its purpose as the temporary branch for the alpha
build of the project and will remain just as it is as a backup for the
future. The primary branch is now master, the migration will include the
separation of code into multiple flies of extensions cpp,h, and ui and
going forward a more modularized approach will be implemented in the
program as the mainwindow.cpp file started to get more and more
cluttered as teh program was growing. The program window will be set in
an aspect ratio of 16:10 and now a more responsive design with layouts
is to be implementd. This commit marks the official migration from main
to master althought master is still incomplete it is now being worked on